Version 1.0 of Hurts
The website currently just handles searches made by the user when filling in the input data. A notification will appear displaying what was returned or if an error occurred. Still need to take that information and display it in a more user friendly manner and create ability to purchase the vehicle to be rented (post request).